Move future annotations into separate module.
